requirements.txt
tensorflow>=2.13
numpy
matplotlib
seaborn
scikit-learn
pillow
kaggle
gradio
pyyaml
tqdm